chore(node): move functional files to `utils` folder
