+++ packages/rolldown/src/utils/create-bundler.ts
@@ -1,8 +1,8 @@
 import { Bundler } from '../binding'
 import { PluginDriver } from '../plugin/plugin-driver'
 import { TreeshakingOptionsSchema } from '../treeshake'
-import { bindingifyInputOptions } from '../options/bindingify-input-options'
-import { bindingifyOutputOptions } from '../options/bindingify-output-options'
+import { bindingifyInputOptions } from './bindingify-input-options'
+import { bindingifyOutputOptions } from './bindingify-output-options'
 import { composeJsPlugins } from './compose-js-plugins'
 import { normalizePluginOption } from './normalize-plugin-option'
 import { initializeParallelPlugins } from './initialize-parallel-plugins'
